Moved feature code to a folder
