+++ app/lib/features/SQLiteConfig/services/database_services.dart
@@ -1,4 +1,4 @@
-import 'package:app/models/task.dart';
+import 'package:app/features/SQLiteConfig/models/task.dart';
 import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:sqflite/sqlite_api.dart';
